requirements.txt
cflib==0.1.21
qtm==2.0.6
numpy==1.24.1
pynput==1.7.6
PyQt5==5.15.7
PyQt5-stubs==5.15.6.0
PythonQwt==0.10.2
qasync==0.23.0
